Fix build errors and warnings
Related to #71

Fix build errors and warnings in the project.

* **GUI/AVBTool.vcxproj**: Change `<SubSystem>` property from `Console` to `Windows` to match the `WinMain` entry point.
* **GUI/AVBTool.cpp**: Add explicit type casting to resolve 'size_t' to 'int' conversion warnings at lines 111 and 133.
* **AVTP/AVTP.vcxproj**: Define `WIN32_LEAN_AND_MEAN` before including any Windows headers to resolve `_WINSOCKAPI_` macro redefinition warnings.
* **gPTP/gPTP.vcxproj**: Define `WIN32_LEAN_AND_MEAN` before including any Windows headers to resolve `_WINSOCKAPI_` macro redefinition warnings.
* **Driver/i210AVBDriver.vcxproj**: Define `WIN32_LEAN_AND_MEAN` before including any Windows headers to resolve `_WINSOCKAPI_` macro redefinition warnings.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/zarfld/AVB-Windows/issues/71?shareId=XXXX-XXXX-XXXX-XXXX).

+++ GUI/AVBTool.cpp
@@ -108,14 +108,14 @@ void AVBTool::stopSynchronization() {
 void AVBTool::synchronize() {
     while (running) {
         const char* syncMessage = "SYNC";
-        int result = sendto(sock, syncMessage, strlen(syncMessage), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
+        int result = sendto(sock, syncMessage, static_cast<int>(strlen(syncMessage)), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
         if (result == SOCKET_ERROR) {
             std::cerr << "Send failed with error: " << WSAGetLastError() << std::endl;
         }
 
         char buffer[1024];
         int serverAddrSize = sizeof(serverAddr);
-        result = recvfrom(sock, buffer, sizeof(buffer), 0, (sockaddr*)&serverAddr, &serverAddrSize);
+        result = recvfrom(sock, buffer, static_cast<int>(sizeof(buffer)), 0, (sockaddr*)&serverAddr, &serverAddrSize);
         if (result == SOCKET_ERROR) {
             std::cerr << "Receive failed with error: " << WSAGetLastError() << std::endl;
         } else {
@@ -130,7 +130,7 @@ void AVBTool::synchronize() {
 void AVBTool::discoverDevices() {
     while (running) {
         const char* discoverMessage = "DISCOVER";
-        int result = sendto(sock, discoverMessage, strlen(discoverMessage), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
+        int result = sendto(sock, discoverMessage, static_cast<int>(strlen(discoverMessage)), 0, (sockaddr*)&serverAddr, sizeof(serverAddr));
         if (result == SOCKET_ERROR) {
             std::cerr << "Send failed with error: " << WSAGetLastError() << std::endl;
             return;
@@ -138,7 +138,7 @@ void AVBTool::discoverDevices() {
 
         char buffer[1024];
         int serverAddrSize = sizeof(serverAddr);
-        result = recvfrom(sock, buffer, sizeof(buffer), 0, (sockaddr*)&serverAddr, &serverAddrSize);
+        result = recvfrom(sock, buffer, static_cast<int>(sizeof(buffer)), 0, (sockaddr*)&serverAddr, &serverAddrSize);
         if (result == SOCKET_ERROR) {
             std::cerr << "Receive failed with error: " << WSAGetLastError() << std::endl;
         } else {
